refactor: align file context module names
