+++ experimental/pi-file-context/src/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./file-quote.js";
+export { default } from "./file-context.js";
